requirements.txt
requests>=2.31.0
supabase>=2.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
psycopg2-binary>=2.9.0
html2text>=2020.1.16
flask>=2.3.0
flask-cors>=4.0.0